% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covax_functions.R
\name{check_location}
\alias{check_location}
\title{Check appointment availability for a time range at one location}
\usage{
check_location(
  location_id,
  start_date = NA,
  end_date = NA,
  vaccineData,
  doseNumber,
  verbose = TRUE
)
}
\arguments{
\item{location_id}{A location's id.}

\item{start_date}{The first date to check for appointments, format YYYY-MM-DD.}

\item{end_date}{The last date to check for appointments, format YYYY-MM-DD.}

\item{vaccineData}{Your personal covaxon API identifier.}

\item{doseNumber}{1 or 2 for first or second dose.}

\item{verbose}{Boolean: would you like updates to the console?}
}
\value{
A tibble containing appointment availabilities.
}
\description{
Check appointment availability for a time range at one location
}
