% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covax_functions.R
\name{check_locations_covax}
\alias{check_locations_covax}
\title{Check appointment availability for a time range at one or more locations}
\usage{
check_locations_covax(locations, end_date = NA, verbose = TRUE)
}
\arguments{
\item{locations}{A set of locations returned from `covaxr::get_locations()`.
This will be a tibble with class `covax` and stored metadata.}

\item{end_date}{The last date to check for appointments, format YYYY-MM-DD.}

\item{verbose}{Boolean: would you like updates to the console?}
}
\value{
The input tibble with additional columns for the date, time, and
number of appointments that were found at each location.
}
\description{
This function checks a
list of locations to see if there are any available COVID-19 vaccination
appointments in a given time range. You can specify the last date to check;
if no last date is given, it looks for appointments within the next 3 days.
}
\details{
It accepts the results of `covaxr::get_locations()`, and you can pipe to it
directly with `%>%`.

It returns the input tibble plus extra columns with information about the
 appointments it found, if any. Details are in nested tibbles in the `available`
 column. If you just want to know how many days have vacancies, or which
 days have vacancies, those values are provided in `num_days` and `days`.
}
