% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covax_functions.R
\name{get_locations}
\alias{get_locations}
\title{Get a list of local vaccination locations}
\usage{
get_locations(
  lat,
  lng,
  fromDate = NA,
  doseNumber,
  vaccineData,
  search_radius = 25
)
}
\arguments{
\item{lat}{Latitude of the location to search from.}

\item{lng}{Longitude of the location to search from.}

\item{fromDate}{Date to begin looking, in format YYYY-MM-DD.}

\item{doseNumber}{Vaccine dose, either 1 or 2.}

\item{vaccineData}{Personal API identifier.}

\item{search_radius}{Range in km to search for locations.}
}
\value{
A tibble with the name, address, locations, and identifiers of all
COVID-19 vaccination clinics matching the input criteria. In addition the
returned object has class \code{covax} that includes metadata about the request--
including the private API access key--that is used by other package functions.
}
\description{
This function returns a data frame with the names, locations, and identifiers
of the COVID-19 vaccination locations that are available for a given
vaccine dose, place, and date.
}
\details{
This function requires a personal API identifier, called \code{vaccineData}, that
the Ontario covax API uses to grant access to the system. You can get your
id by logging into the \href{https://vaccine.covaxonbooking.ca/}{provincial booking system}
using Chrome and opening the developer console by pressing \code{Ctrl-Shift-J}
(on a Windows machine--consult Google for advice on other machines)
\itemize{
\item https://vaccine.covaxonbooking.ca/
}
}
\examples{
\dontrun{
vax_data <- "YOUR_ONLINE_API_IDENTIFIER"
apts <- covaxr::get_locations(lat = 43.692556,
                              lng =  -79.377056,
                              fromDate = as.character(Sys.Date()),
                              doseNumber = 1,
                              vaccineData = vax_data) \%>\%
        covaxr::check_locations_covax(end_date = "2021-09-01")
}
}
