% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covax_functions.R
\name{map_locations}
\alias{map_locations}
\title{Map COVID-19 vaccination locations}
\usage{
map_locations(locs)
}
\arguments{
\item{locs}{A tibble containing COVID-19 vaccination locations obtained
through `covaxr::get_locations()`.}
}
\value{
A **leaflet** map.
}
\description{
This function makes an interactive map of local COVID-19 vaccination stations
using the **leaflet** package. It accepts input from
}
\examples{
\dontrun{
locations <- covaxr::get_locations(lat = 43.692556,
                      lng =  -79.377056,
                      fromDate = as.character(Sys.Date()),
                      doseNumber = 1,
                      vaccineData = vax_data)

covaxr::map_locations(locations)
}
}
